requirements.txt
python-telegram-bot
gspread
oauth2client
gspread-formatting
fastapi
uvicorn
requests
